rtx: start splitting ray tracing into functional modules
move model management into vk_ray_model
